Specify dataset as command option

+++ launch_docker.sh
@@ -1,18 +1,7 @@
 #!/bin/bash
 
-if [[ -z "${FLOWNET2_CHECKPOINTS_DIR}" ]]; then
-    echo "Environment variable 'FLOWNET2_CHECKPOINTS_DIR' does not exist."
-    echo "Please set it ~/.bashrc using: 'export FLOWNET2_CHECKPOINTS_DIR=/path/to/FLOWNET2_CHECKPOINTS_DIR'"
-    exit 1
-fi
-
-if [[ -z "${DATASET_DIR}" ]]; then
-    echo "Environment variable 'DATASET_DIR' does not exist."
-    echo "Please set it in ~/.bashrc using: 'export DATASET_DIR=/path/to/DATASET_DIR'"
-    exit 1
-fi
-
 COMMAND="/bin/bash"
+dataset_dir=${DATASET_DIR}
 
 while test $# -gt 0
 do
@@ -21,16 +10,31 @@ do
             ;;
         --run-convert) COMMAND="./run.sh --convert"
             ;;
-        --install) COMMAND="./install.sh"
-            ;;
+		--dataset)
+			dataset_dir="$2"
+            echo $dataset_dir
+			shift
+			;;
     esac
     shift
 done
 
+if [[ -z "${FLOWNET2_CHECKPOINTS_DIR}" ]]; then
+    echo "Environment variable 'FLOWNET2_CHECKPOINTS_DIR' does not exist."
+    echo "Please set it ~/.bashrc using: 'export FLOWNET2_CHECKPOINTS_DIR=/path/to/FLOWNET2_CHECKPOINTS_DIR'"
+    exit 1
+fi
+
+if [[ -z "${dataset_dir}" ]]; then
+    echo "Environment variable 'DATASET_DIR' does not exist."
+    echo "Please set it in ~/.bashrc using: 'export DATASET_DIR=/path/to/DATASET_DIR'"
+    exit 1
+fi
+
 docker run  --gpus all --rm -ti \
             --volume=$(pwd):/flownet2-pytorch:rw \
             -v $FLOWNET2_CHECKPOINTS_DIR:/data/flownet2-checkpoints \
-            -v $DATASET_DIR:/data/frames \
+            -v $dataset_dir:/data/frames \
             --workdir=/flownet2-pytorch \
             --ipc=host $USER/flownet2:latest \
             $COMMAND
